requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
awsebcli==3.20.10
blessed==1.20.0
blinker==1.4
botocore==1.31.85
branca==0.8.0
cachetools==5.5.0
cement==2.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.5
colorama==0.4.3
command-not-found==0.3
cryptography==3.4.8
dataclasses-json==0.6.7
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
exceptiongroup==1.2.2
faiss-cpu==1.9.0
filelock==3.16.1
folium==0.18.0
frozenlist==1.5.0
fsspec==2024.10.0
gguf @ file:///home/sushidolly/BitNet/3rdparty/llama.cpp/gguf-py
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.8
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.5.0
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.15
langchain-ollama==0.2.0
langchain-openai==0.2.5
langchain-text-splitters==0.3.1
langsmith==0.1.137
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
narwhals==1.12.1
netifaces==0.11.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.0
ollama==0.3.3
openai==1.53.0
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pathspec==0.10.1
pillow==10.4.0
propcache==0.2.0
protobuf==4.25.5
pyarrow==18.0.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
PyMySQL==1.1.1
pyparsing==2.4.7
PyPDF2==3.0.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==5.4.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
rpds-py==0.20.0
safetensors==0.4.5
SecretStorage==3.3.1
semantic-version==2.8.5
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
streamlit==1.39.0
streamlit-chat==0.1.1
streamlit_folium==0.23.1
sympy==1.13.3
systemd-python==234
tenacity==8.5.0
termcolor==1.1.0
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
torch==2.2.2+cpu
tornado==6.4.1
tqdm==4.66.5
transformers==4.46.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.20
wadllib==1.3.6
watchdog==5.0.3
wcwidth==0.1.9
xyzservices==2024.9.0
yarl==1.17.0
zipp==1.0.0
